stok controller stok getir metodu düzenlendi. sok model düzenlendi.

+++ AtikJr/Models/StokModel.cs
@@ -30,6 +30,6 @@
         /// <summary>
         /// stok sayısı
         /// </summary>
-        public float STOK_MİKTAR { get; set; }
+        public int STOK_MİKTAR { get; set; }
     }
 }
